pyproject.toml
[tool.poetry]
name = "sunbeam"
version = "0.1.0"
description = "UBC Solar's data pipeline: logic, distribution, storage, and infrastructure"
authors = ["UBC Solar Strategy <strategy@ubcsolar.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.8.2"
toml = "^0.10.2"
prefect = "^2.20.2"
prefect-github = "^0.2.7"
python-dotenv = "^1.0.1"
pymongo = "^4.8.0"
tqdm = "^4.66.5"
ubc-solar-data-tools = {git = "https://github.com/UBC-Solar/data_tools", rev = "sunbeam_utils"}
dill = "^0.3.9"

[tool.poetry.group.external.dependencies]
flask = {version = "^3.0.3", extras = ["async"]}

[tool.poetry.group.external]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
